feat: Increase pagination limit and use data
