+++ backend/config/default.json
@@ -6,8 +6,8 @@
     "http://localhost:3030"
   ],
   "paginate": {
-    "default": 10,
-    "max": 50
+    "default": 999,
+    "max": 999
   },
   "mongodb": "mongodb://127.0.0.1:27017/backend"
 }
