+++ backend/src/services/boards/boards.hooks.ts
@@ -1,7 +1,7 @@
 import type { HookContext } from '../../declarations';
 
 export const afterBoardRemove = async (context: HookContext) => {
-  if (context.id !== undefined) {
-    await context.app.service('lists').remove(null, { query: { boardId: context.id.toString() } });
+  if (context.result?._id) {
+    await context.app.service('lists').remove(null, { query: { boardId: context.result._id.toString() } });
   }
 };
