+++ backend/src/services/lists/lists.hooks.ts
@@ -1,7 +1,7 @@
 import type { HookContext } from '../../declarations';
 
 export const afterListRemove = async (context: HookContext) => {
-  if (context.id !== undefined) {
-    await context.app.service('cards').remove(null, { query: { listId: context.id.toString() } });
+  if (context.result?._id) {
+    await context.app.service('cards').remove(null, { query: { listId: context.result._id.toString() } });
   }
 };
